requirements.txt
streamlit==0.81.0
numpy==1.20.1
pandas==1.2.1
pillow==8.1.0